+++ core/src/main/scala/org/broadinstitute/dsde/rawls/util/AttributeSupport.scala
@@ -1,0 +1,134 @@
+package org.broadinstitute.dsde.rawls.util
+
+import akka.http.scaladsl.model.StatusCodes
+import org.broadinstitute.dsde.rawls.RawlsExceptionWithErrorReport
+import org.broadinstitute.dsde.rawls.model.Attributable.AttributeMap
+import org.broadinstitute.dsde.rawls.model.AttributeUpdateOperations.{AddListMember, AddUpdateAttribute, AttributeUpdateOperation, CreateAttributeEntityReferenceList, CreateAttributeValueList, RemoveAttribute, RemoveListMember}
+import org.broadinstitute.dsde.rawls.model.{Attributable, AttributeEntityReference, AttributeEntityReferenceEmptyList, AttributeEntityReferenceList, AttributeName, AttributeNull, AttributeValue, AttributeValueEmptyList, AttributeValueList, ErrorReport, MethodConfiguration}
+import org.broadinstitute.dsde.rawls.workspace.{AttributeNotFoundException, AttributeUpdateOperationException}
+
+import scala.concurrent.Future
+
+trait AttributeSupport {
+
+  // note: success is indicated by  Map.empty
+  def attributeNamespaceCheck(attributeNames: Iterable[AttributeName]): Map[String, String] = {
+    val namespaces = attributeNames.map(_.namespace).toSet
+
+    // no one can modify attributes with invalid namespaces
+    val invalidNamespaces = namespaces -- AttributeName.validNamespaces
+    invalidNamespaces.map { ns => ns -> s"Invalid attribute namespace $ns" }.toMap
+  }
+
+  def withAttributeNamespaceCheck[T](attributeNames: Iterable[AttributeName])(op: => T): T = {
+    val errors = attributeNamespaceCheck(attributeNames)
+    if (errors.isEmpty) op
+    else {
+      val reasons = errors.values.mkString(", ")
+      val err = ErrorReport(statusCode = StatusCodes.Forbidden, message = s"Attribute namespace validation failed: [$reasons]")
+      throw new RawlsExceptionWithErrorReport(errorReport = err)
+    }
+  }
+
+  def withAttributeNamespaceCheck[T](hasAttributes: Attributable)(op: => Future[T]): Future[T] =
+    withAttributeNamespaceCheck(hasAttributes.attributes.keys)(op)
+
+  def withAttributeNamespaceCheck[T](methodConfiguration: MethodConfiguration)(op: => Future[T]): Future[T] = {
+    // TODO: this duplicates expression parsing, the canonical way to do this.  Use that instead?
+    // valid method configuration outputs are either in the format this.attrname or workspace.attrname
+    // invalid (unparseable) will be caught by expression parsing instead
+    val attrNames = methodConfiguration.outputs map { case (_, attr) => AttributeName.fromDelimitedName(attr.value.split('.').last) }
+    withAttributeNamespaceCheck(attrNames)(op)
+  }
+
+  def applyAttributeUpdateOperations(attributable: Attributable, operations: Seq[AttributeUpdateOperation]): AttributeMap = {
+    operations.foldLeft(attributable.attributes) { (startingAttributes, operation) =>
+
+      operation match {
+        case AddUpdateAttribute(attributeName, attribute) => startingAttributes + (attributeName -> attribute)
+
+        case RemoveAttribute(attributeName) => startingAttributes - attributeName
+
+        case CreateAttributeEntityReferenceList(attributeName) =>
+          if( startingAttributes.contains(attributeName) ) { //non-destructive
+            startingAttributes
+          } else {
+            startingAttributes + (attributeName -> AttributeEntityReferenceEmptyList)
+          }
+
+        case CreateAttributeValueList(attributeName) =>
+          if( startingAttributes.contains(attributeName) ) { //non-destructive
+            startingAttributes
+          } else {
+            startingAttributes + (attributeName -> AttributeValueEmptyList)
+          }
+
+        case AddListMember(attributeListName, newMember) =>
+          startingAttributes.get(attributeListName) match {
+            case Some(AttributeValueEmptyList) =>
+              newMember match {
+                case AttributeNull =>
+                  startingAttributes
+                case newMember: AttributeValue =>
+                  startingAttributes + (attributeListName -> AttributeValueList(Seq(newMember)))
+                case newMember: AttributeEntityReference =>
+                  throw new AttributeUpdateOperationException("Cannot add non-value to list of values.")
+                case _ => throw new AttributeUpdateOperationException("Cannot create list with that type.")
+              }
+
+            case Some(AttributeEntityReferenceEmptyList) =>
+              newMember match {
+                case AttributeNull =>
+                  startingAttributes
+                case newMember: AttributeEntityReference =>
+                  startingAttributes + (attributeListName -> AttributeEntityReferenceList(Seq(newMember)))
+                case newMember: AttributeValue =>
+                  throw new AttributeUpdateOperationException("Cannot add non-reference to list of references.")
+                case _ => throw new AttributeUpdateOperationException("Cannot create list with that type.")
+              }
+
+            case Some(l: AttributeValueList) =>
+              newMember match {
+                case AttributeNull =>
+                  startingAttributes
+                case newMember: AttributeValue =>
+                  startingAttributes + (attributeListName -> AttributeValueList(l.list :+ newMember))
+                case _ => throw new AttributeUpdateOperationException("Cannot add non-value to list of values.")
+              }
+
+            case Some(l: AttributeEntityReferenceList) =>
+              newMember match {
+                case AttributeNull =>
+                  startingAttributes
+                case newMember: AttributeEntityReference =>
+                  startingAttributes + (attributeListName -> AttributeEntityReferenceList(l.list :+ newMember))
+                case _ => throw new AttributeUpdateOperationException("Cannot add non-reference to list of references.")
+              }
+
+            case None =>
+              newMember match {
+                case AttributeNull =>
+                  throw new AttributeUpdateOperationException("Cannot use AttributeNull to create empty list. Use CreateEmpty[Ref|Val]List instead.")
+                case newMember: AttributeValue =>
+                  startingAttributes + (attributeListName -> AttributeValueList(Seq(newMember)))
+                case newMember: AttributeEntityReference =>
+                  startingAttributes + (attributeListName -> AttributeEntityReferenceList(Seq(newMember)))
+                case _ => throw new AttributeUpdateOperationException("Cannot create list with that type.")
+              }
+
+            case Some(_) => throw new AttributeUpdateOperationException(s"$attributeListName of ${attributable.briefName} is not a list")
+          }
+
+        case RemoveListMember(attributeListName, removeMember) =>
+          startingAttributes.get(attributeListName) match {
+            case Some(l: AttributeValueList) =>
+              startingAttributes + (attributeListName -> AttributeValueList(l.list.filterNot(_ == removeMember)))
+            case Some(l: AttributeEntityReferenceList) =>
+              startingAttributes + (attributeListName -> AttributeEntityReferenceList(l.list.filterNot(_ == removeMember)))
+            case None => throw new AttributeNotFoundException(s"$attributeListName of ${attributable.briefName} does not exist")
+            case Some(_) => throw new AttributeUpdateOperationException(s"$attributeListName of ${attributable.briefName} is not a list")
+          }
+      }
+    }
+  }
+}
