+++ core/src/main/scala/org/broadinstitute/dsde/rawls/util/EntitySupport.scala
@@ -1,0 +1,51 @@
+package org.broadinstitute.dsde.rawls.util
+
+import akka.http.scaladsl.model.StatusCodes
+import org.broadinstitute.dsde.rawls.{RawlsException, RawlsExceptionWithErrorReport}
+import org.broadinstitute.dsde.rawls.dataaccess.slick.{DataAccess, EntityRecord, ReadAction, ReadWriteAction}
+import org.broadinstitute.dsde.rawls.dataaccess.{SlickDataSource, SlickWorkspaceContext}
+import org.broadinstitute.dsde.rawls.model.{AttributeEntityReference, Entity, ErrorReport}
+import org.broadinstitute.dsde.rawls.webservice.PerRequest.PerRequestMessage
+
+import scala.concurrent.ExecutionContext
+import scala.util.{Failure, Success, Try}
+
+trait EntitySupport {
+  implicit protected val executionContext: ExecutionContext
+  protected val dataSource: SlickDataSource
+
+  import dataSource.dataAccess.driver.api._
+
+  //Finds a single entity record in the db.
+  def withSingleEntityRec(entityType: String, entityName: String, workspaceContext: SlickWorkspaceContext, dataAccess: DataAccess)(op: (Seq[EntityRecord]) => ReadWriteAction[PerRequestMessage]): ReadWriteAction[PerRequestMessage] = {
+    val entityRec = dataAccess.entityQuery.findEntityByName(workspaceContext.workspaceId, entityType, entityName).result
+    entityRec flatMap { entities =>
+      if (entities.isEmpty) {
+        DBIO.failed(new RawlsExceptionWithErrorReport(errorReport = ErrorReport(StatusCodes.NotFound, s"No entity of type ${entityType} named ${entityName} exists in this workspace.")))
+      } else if (entities.size == 1) {
+        op(entities)
+      } else {
+        DBIO.failed(new RawlsExceptionWithErrorReport(errorReport = ErrorReport(StatusCodes.NotFound, s"More than one entity of type ${entityType} named ${entityName} exists in this workspace?!")))
+      }
+    }
+  }
+
+  def withAllEntities[T](workspaceContext: SlickWorkspaceContext, dataAccess: DataAccess, entities: Seq[AttributeEntityReference])(op: (Seq[Entity]) => ReadWriteAction[T]): ReadWriteAction[T] = {
+    val entityActions: Seq[ReadAction[Try[Entity]]] = entities map { e =>
+      dataAccess.entityQuery.get(workspaceContext, e.entityType, e.entityName) map {
+        case None => Failure(new RawlsException(s"${e.entityType} ${e.entityName} does not exist in ${workspaceContext.workspace.toWorkspaceName}"))
+        case Some(entity) => Success(entity)
+      }
+    }
+
+    DBIO.sequence(entityActions) flatMap { entityTries =>
+      val failures = entityTries.collect { case Failure(y) => y.getMessage }
+      if (failures.isEmpty) op(entityTries collect { case Success(e) => e })
+      else {
+        val err = ErrorReport(statusCode = StatusCodes.BadRequest, message = (Seq("Entities were not found:") ++ failures) mkString System.lineSeparator())
+        DBIO.failed(new RawlsExceptionWithErrorReport(err))
+      }
+    }
+  }
+
+}
